requirements.txt
aiofiles==24.1.0
asyncio==3.4.3
attrs==24.2.0
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
exceptiongroup==1.2.2
firebase-admin==6.5.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.1
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
msgpack==1.1.0
packaging==24.1
pkgutil_resolve_name==1.3.10
pluggy==1.5.0
priority==2.0.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.1.4
pytest==8.3.2
python-dotenv==1.0.1
PyYAML==6.0.2
Quart==0.19.6
quart-cors==0.7.0
quart-openapi==1.7.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
six==1.16.0
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.4
wsproto==1.2.0
zipp==3.19.2
